Fix receipt.

+++ app/Http/Controllers/Frontend/Sale/SaleController.php
@@ -20,7 +20,8 @@ use DB;
 
 class SaleController extends Controller
 {
-    public function index(){
+    public function index()
+    {
         $total  = 0;
         $orders = Order::where(Db::raw('date(created_at)'), date('Y-m-d'))
                   ->where('user_id', Auth::user()->id)
@@ -35,7 +36,8 @@ class SaleController extends Controller
         return view('frontend.user.sale.daily', compact('total', 'orders'));
     }
 
-    public function monthly(){
+    public function monthly()
+    {
         $total  = 0;
         $date   = date('F Y');
         $orders = Order::whereBetween(Db::raw('date(created_at)'), [date('Y-m-01'), date('Y-m-31')])
@@ -51,8 +53,14 @@ class SaleController extends Controller
         return view('frontend.user.sale.monthly', compact('total', 'orders', 'date'));
     }
 
+    # SAVE FUNCTION
     public function save(Request $request){
     	$arr 	 = json_decode($request->orders);
+
+    	// return json_encode($request->orders);
+
+    	// "[{\"id\":\"2\",\"code\":\"Coke\",\"price\":\"30.00\",\"qty\":\"1\",\"size\":\"Short Order\"}]"
+
         $p_avail = 0; //available product
 
         if(empty($request->transaction_no))
@@ -95,7 +103,7 @@ class SaleController extends Controller
             //
             for($i = 0; $i < count($arr); $i++)
             {
-                $size = ProductSize::where('size', $arr[$i]->size)->first();
+                $size = ProductSize::where('product_id', $arr[$i]->id)->first();
 
                 $list = OrderList::updateOrCreate(
                     [
@@ -146,11 +154,13 @@ class SaleController extends Controller
 
             $this->notification();
 
-            $order = Order::with(
+            $orderList = Order::with(
                     'order_list', 'order_list.product', 'order_list.product_size'
                 )->where('id', $order->id)->first();
 
-            return json_encode(['status' => 'success', 'order' => $order]);
+            // $order_list = OrderList::with('product', 'product_size')->where('order_id', $order->id)->get();
+
+            return json_encode(['status' => 'success', 'order' => $orderList]);
         }
         else
         {
@@ -159,10 +169,12 @@ class SaleController extends Controller
 
     	return ['status' => 'failed'];
     }
+    # END SAVE FUNCTION
 
-    public function charge_save(Request $request) {
-        $order = Order::with('order_list', 'order_list.product', 'order_list.product_size','table')
-                ->where('transaction_no', $request->transaction_no)
+
+    public function charge_save(Request $request)
+    {
+        $order = Order::where('transaction_no', $request->transaction_no)
                 ->first();
 
         $order->cash        = $request->cash;
@@ -172,9 +184,12 @@ class SaleController extends Controller
         $order->vat         = $request->vat;
         $order->total       = $request->amount_due;
         $order->status      = 'Paid';
-        $order->save();
 
-        return json_encode(['status' => 'success', 'order' => $order]);
+        if($order->save()) {
+            $order_list = OrderList::with('product', 'product_size', 'order')->whereOrderId($order->id)->get();
+
+            return json_encode(['status' => 'success', 'order' => $order_list]);
+        }
     }
 
     public function product_availability($product_id, $quantity)
@@ -196,21 +211,21 @@ class SaleController extends Controller
                 if($inventory->physical_quantity == 'Mass')
                 {
                     $stock_qty = new Mass($inventory->stock, $inventory->unit_type);
-
+        
                     $req_qty   = new Mass(($quantity * $inventory->pivot->quantity), $inventory->pivot->unit_type);
-
+        
                     $qty_left  = $stock_qty->subtract($req_qty);
-
+        
                     $stock_dec = $qty_left->toUnit($inventory->unit_type);
                 }
                 elseif($inventory->physical_quantity == 'Volume')
                 {
                     $stock_qty = new Volume($inventory->stock, $inventory->unit_type);
-
+        
                     $req_qty   = new Volume(($quantity * $inventory->pivot->quantity), $inventory->pivot->unit_type);
-
+        
                     $qty_left  = $stock_qty->subtract($req_qty);
-
+        
                     $stock_dec = $qty_left->toUnit($inventory->unit_type);
                 }
                 else
@@ -229,7 +244,8 @@ class SaleController extends Controller
         return $status;
     }
 
-    public function available_table(){
+    public function available_table()
+    {
         $available = [];
         $pendings = Order::select('table_no')
                 ->where('status', 'Unpaid')
@@ -241,7 +257,8 @@ class SaleController extends Controller
         return json_encode($tables);
     }
 
-    public function unpaid(){
+    public function unpaid()
+    {
         $tables = Order::select('table_no')
                 ->where('status', 'Unpaid')
                 ->where('type', 'Dine-in')
@@ -251,24 +268,27 @@ class SaleController extends Controller
         return $tables;
     }
 
-    public function get_order($table) {
+    public function get_order($table)
+    {
         $order = Order::with(
                     'order_list', 'order_list.product', 'order_list.product_size', 'table'
-                )->where('status', 'Unpaid')
-                ->where('type', 'Dine-in')
-                ->whereBetween('created_at', [Date('Y-m-d 00:00:00'), Date('Y-m-d 23:59:59')])
-                ->where('table_no', $table)
-                ->first();
+            )->where('status', 'Unpaid')
+            ->where('type', 'Dine-in')
+            ->whereBetween('created_at', [Date('Y-m-d 00:00:00'), Date('Y-m-d 23:59:59')])
+            ->where('table_no', $table)
+            ->first();
 
         return json_encode(['order' => $order]);
     }
 
-    public function get_order_list($transaction_no) {
-        $order = Order::with('order_list', 'order_list.product', 'order_list.product_size', 'table')
-                ->where('transaction_no', $transaction_no)
-                ->first();
+    public function get_order_list($transaction_no)
+    {
+        $order = Order::where('transaction_no', $transaction_no)->first();
+        $order_list = OrderList::with(  'product', 'product_size', 'order.table')
+            ->where('order_id', $order->id)
+            ->get();
 
-        return $order;
+        return $order_list;
     }
 
     public function cancel_order(Request $request)
@@ -282,19 +302,25 @@ class SaleController extends Controller
             $stock = 0;
 
             foreach($list->product_size->ingredients as $ingredient) {
-                $stock = $list->quantity * $ingredient->pivot->quantity;
-                $ingredient_stock = new Mass($ingredient->stock, $ingredient->unit_type);
-                $overall_stock = $ingredient_stock->add(new Mass($stock, $ingredient->pivot->unit_type));
-                $ingredient->stock = $overall_stock;
-                $ingredient->save();
-
-                $list->delete();
+                if (($ingredient->physical_quantity == 'Mass') || ($ingredient->physical_quantity == 'Volume')) {
+                    $stock = $list->quantity * $ingredient->pivot->quantity; // Used ingredient Stock
+                    $ingredient_stock = new Mass($ingredient->stock, $ingredient->unit_type); // Actual Stock
+                    $overall_stock = $ingredient_stock->add(new Mass($stock, $ingredient->pivot->unit_type));
+                    $ingredient->stock = $overall_stock;
+                    if($ingredient->save())
+                        $list->delete();
+                } else {
+                    $stock = $list->quantity * $ingredient->pivot->quantity;
+                    $ingredient->stock = $ingredient->stock + $stock;
+                    if($ingredient->save())
+                        $list->delete();
+                }
             }
         }
 
         $order = Order::where('transaction_no', $request->transaction_no)->first();
 
-        if(count($order->order_list) !== 0)
+        if(count($order->order_list) == 0)
         {
             $order->status = 'Cancelled';
             $order->save();
@@ -303,7 +329,47 @@ class SaleController extends Controller
         return json_encode(['status' => $order->status]);
     }
 
-    public function notification(){
+    /*public function check_cancel_order()
+    {
+        $order = Order::whereTransactionNo('18-02229845')->first();
+        // dd($order);
+        $order_list = OrderList::whereOrderId($order->id)->with('product_size')->get();
+        // dd($order_list);
+
+        foreach($order_list as $list)
+        {
+            $stock = 0;
+
+            foreach($list->product_size->ingredients as $ingredient) {
+                if (($ingredient->physical_quantity == 'Mass') || ($ingredient->physical_quantity == 'Volume')) {
+                    $stock = $list->quantity * $ingredient->pivot->quantity; // Used ingredient Stock
+                    $ingredient_stock = new Mass($ingredient->stock, $ingredient->unit_type); // Actual Stock
+                    $overall_stock = $ingredient_stock->add(new Mass($stock, $ingredient->pivot->unit_type));
+                    $ingredient->stock = $overall_stock;
+                    if($ingredient->save())
+                        $list->delete();
+                } else {
+                    $stock = $list->quantity * $ingredient->pivot->quantity;
+                    $ingredient->stock = $ingredient->stock + $stock;
+                    if($ingredient->save())
+                        $list->delete();
+                }
+            }
+        }
+
+        $order = Order::where('transaction_no', '18-02613449')->first();
+
+        if(count($order->order_list) == 0)
+        {
+            $order->status = 'Cancelled';
+            $order->save();
+        }
+
+        return json_encode(['status' => $order->status]);
+    }*/
+
+    public function notification()
+    {
         $inventories = Inventory::whereRaw('stock < reorder_level')->get();
 
         foreach ($inventories as $inventory) {
